+++ src/Microsoft.DotNet.Darc/src/Darc/Operations/AddSubscriptionOperation.cs
@@ -31,7 +31,7 @@ namespace Microsoft.DotNet.Darc.Operations
         /// <param name="options"></param>
         public override async Task<int> ExecuteAsync()
         {
-            DarcSettings darcSettings = LocalCommands.GetSettings(_options, Logger);
+            DarcSettings darcSettings = LocalSettings.GetDarcSettings(_options, Logger);
             // No need to set up a git type or PAT here.
             Remote remote = new Remote(darcSettings, Logger);
