+++ src/Microsoft.DotNet.Darc/src/Darc/Operations/GetChannelsOperation.cs
@@ -29,7 +29,7 @@ namespace Microsoft.DotNet.Darc.Operations
         {
             try
             {
-                DarcSettings darcSettings = LocalCommands.GetSettings(_options, Logger);
+                DarcSettings darcSettings = LocalSettings.GetDarcSettings(_options, Logger);
                 // No need to set up a git type or PAT here.
                 Remote remote = new Remote(darcSettings, Logger);
