+++ src/Microsoft.DotNet.Darc/src/Darc/Operations/GetDependenciesOperation.cs
@@ -5,6 +5,7 @@
 using Microsoft.DotNet.Darc.Helpers;
 using Microsoft.DotNet.Darc.Options;
 using Microsoft.DotNet.DarcLib;
+using Microsoft.DotNet.DarcLib.Helpers;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
@@ -25,7 +26,7 @@ namespace Microsoft.DotNet.Darc.Operations
 
         public override async Task<int> ExecuteAsync()
         {
-            Local local = new Local(LocalCommands.GetGitDir(Logger), Logger);
+            Local local = new Local(LocalHelpers.GetGitDir(Logger), Logger);
 
             try
             {
